mummy: Conditionally push /linkresource args into the response file
The new modular structure of VTK results in 85 dlls by default,
which can easily result in command lines well over the 8K limit
if all of their full paths are listed on the command line.

Invent a way for the /linkresource args to be in the response
file to avoid "command line too long" problems.

If the caller of ADD_CSHARP_* sets MUMMY_ADD_CSHARP_CONFIG then
the string "${CMAKE_CFG_INTDIR}" in any linkresource reference
(i.e. "$(OutDir)") will be replaced by ${MUMMY_ADD_CSHARP_CONFIG}.

Unfortunately, for projects that need to use this feature,
that means setting the MUMMY_ADD_CSHARP_CONFIG value at CMake
configure time rather than at build time.

This is the quick way of fixing it, so we can use it without
completely refactoring this to run through a generated script
instead.

If instead, we ran it through a generated CMake script, we could
pass in -DMUMMY_ADD_CSHARP_CONFIG=${CMAKE_CFG_INTDIR} and then we
could do the replacement at build time when that script gets run.

But that refactoring would take another couple of days to get it
working properly... Save for later.